db: Combine files that deal with db.UserStore
We had methods spread across multiple files which makes it difficult to
find them all.
