+++ internal/db/users.go
@@ -2,8 +2,11 @@ package db
 
 import (
 	"context"
+	"crypto/rand"
 	"database/sql"
+	"encoding/base64"
 	"fmt"
+	"strconv"
 	"sync"
 	"unicode/utf8"
 
@@ -11,14 +14,17 @@ import (
 	"github.com/keegancsmith/sqlf"
 	"github.com/lib/pq"
 	"github.com/pkg/errors"
+	"golang.org/x/crypto/bcrypt"
 
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/db/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
+	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/db/globalstatedb"
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
+	"github.com/sourcegraph/sourcegraph/internal/randstring"
 	"github.com/sourcegraph/sourcegraph/internal/timeutil"
 	"github.com/sourcegraph/sourcegraph/internal/trace"
 	"github.com/sourcegraph/sourcegraph/internal/types"
@@ -838,4 +844,224 @@ func (u *UserStore) getBySQL(ctx context.Context, query *sqlf.Query) ([]*types.U
 	}
 
 	return users, nil
+}
+
+func (u *UserStore) IsPassword(ctx context.Context, id int32, password string) (bool, error) {
+	u.ensureStore()
+
+	var passwd sql.NullString
+	if err := u.QueryRow(ctx, sqlf.Sprintf("SELECT passwd FROM users WHERE deleted_at IS NULL AND id=%s", id)).Scan(&passwd); err != nil {
+		return false, err
+	}
+	if !passwd.Valid {
+		return false, nil
+	}
+	return validPassword(passwd.String, password), nil
+}
+
+var (
+	passwordResetRateLimit    = "1 minute"
+	ErrPasswordResetRateLimit = errors.New("password reset rate limit reached")
+)
+
+func (u *UserStore) RenewPasswordResetCode(ctx context.Context, id int32) (string, error) {
+	u.ensureStore()
+
+	if _, err := u.GetByID(ctx, id); err != nil {
+		return "", err
+	}
+	var b [40]byte
+	if _, err := rand.Read(b[:]); err != nil {
+		return "", err
+	}
+	code := base64.StdEncoding.EncodeToString(b[:])
+	res, err := u.ExecResult(ctx, sqlf.Sprintf("UPDATE users SET passwd_reset_code=%s, passwd_reset_time=now() WHERE id=%s AND (passwd_reset_time IS NULL OR passwd_reset_time + interval '"+passwordResetRateLimit+"' < now())", code, id))
+	if err != nil {
+		return "", err
+	}
+	affected, err := res.RowsAffected()
+	if err != nil {
+		return "", err
+	}
+	if affected == 0 {
+		return "", ErrPasswordResetRateLimit
+	}
+
+	return code, nil
+}
+
+// SetPassword sets the user's password given a new password and a password reset code
+func (u *UserStore) SetPassword(ctx context.Context, id int32, resetCode, newPassword string) (bool, error) {
+	u.ensureStore()
+
+	// 🚨 SECURITY: no empty passwords
+	if newPassword == "" {
+		return false, errors.New("new password was empty")
+	}
+
+	resetLinkExpiryDuration := conf.AuthPasswordResetLinkExpiry()
+
+	// 🚨 SECURITY: check resetCode against what's in the DB and that it's not expired
+	r := u.QueryRow(ctx, sqlf.Sprintf("SELECT count(*) FROM users WHERE id=%s AND deleted_at IS NULL AND passwd_reset_code=%s AND passwd_reset_time + interval '"+strconv.Itoa(resetLinkExpiryDuration)+" seconds' > now()", id, resetCode))
+
+	var ct int
+	if err := r.Scan(&ct); err != nil {
+		return false, err
+	}
+	if ct > 1 {
+		return false, fmt.Errorf("illegal state: found more than one user matching ID %d", id)
+	}
+	if ct == 0 {
+		return false, nil
+	}
+	passwd, err := hashPassword(newPassword)
+	if err != nil {
+		return false, err
+	}
+	// 🚨 SECURITY: set the new password and clear the reset code and expiry so the same code can't be reused.
+	if err := u.Exec(ctx, sqlf.Sprintf("UPDATE users SET passwd_reset_code=NULL, passwd_reset_time=NULL, passwd=%s WHERE id=%s", passwd, id)); err != nil {
+		return false, err
+	}
+
+	return true, nil
+}
+
+func (u *UserStore) DeletePasswordResetCode(ctx context.Context, id int32) error {
+	u.ensureStore()
+
+	err := u.Exec(ctx, sqlf.Sprintf("UPDATE users SET passwd_reset_code=NULL, passwd_reset_time=NULL WHERE id=%s", id))
+	return err
+}
+
+// UpdatePassword updates a user's password given the current password.
+func (u *UserStore) UpdatePassword(ctx context.Context, id int32, oldPassword, newPassword string) error {
+	u.ensureStore()
+
+	// 🚨 SECURITY: No empty passwords.
+	if oldPassword == "" {
+		return errors.New("old password was empty")
+	}
+	if newPassword == "" {
+		return errors.New("new password was empty")
+	}
+	// 🚨 SECURITY: Make sure the caller provided the correct old password.
+	if ok, err := u.IsPassword(ctx, id, oldPassword); err != nil {
+		return err
+	} else if !ok {
+		return errors.New("wrong old password")
+	}
+
+	if err := CheckPasswordLength(newPassword); err != nil {
+		return err
+	}
+
+	passwd, err := hashPassword(newPassword)
+	if err != nil {
+		return err
+	}
+	// 🚨 SECURITY: Set the new password
+	if err := u.Exec(ctx, sqlf.Sprintf("UPDATE users SET passwd_reset_code=NULL, passwd_reset_time=NULL, passwd=%s WHERE id=%s", passwd, id)); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// RandomizePasswordAndClearPasswordResetRateLimit overwrites a user's password with a hard-to-guess
+// random password and clears the password reset rate limit. It is intended to be used by site admins,
+// who can subsequently generate a new password reset code for the user (in case the user has locked
+// themselves out, or in case the site admin wants to initiate a password reset).
+//
+// A randomized password is used (instead of an empty password) to avoid bugs where an empty password
+// is considered to be no password. The random password is expected to be irretrievable.
+func (u *UserStore) RandomizePasswordAndClearPasswordResetRateLimit(ctx context.Context, id int32) error {
+	u.ensureStore()
+
+	passwd, err := hashPassword(randstring.NewLen(36))
+	if err != nil {
+		return err
+	}
+	// 🚨 SECURITY: Set the new random password and clear the reset code/expiry, so the old code
+	// can't be reused, and so a new valid reset code can be generated afterward.
+	err = u.Exec(ctx, sqlf.Sprintf("UPDATE users SET passwd_reset_code=NULL, passwd_reset_time=NULL, passwd=%s WHERE id=%s", passwd, id))
+	return err
+}
+
+func hashPassword(password string) (sql.NullString, error) {
+	if dbtesting.MockHashPassword != nil {
+		return dbtesting.MockHashPassword(password)
+	}
+	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
+	if err != nil {
+		return sql.NullString{}, err
+	}
+	return sql.NullString{Valid: true, String: string(hash)}, nil
+}
+
+func validPassword(hash, password string) bool {
+	if dbtesting.MockValidPassword != nil {
+		return dbtesting.MockValidPassword(hash, password)
+	}
+	return bcrypt.CompareHashAndPassword([]byte(hash), []byte(password)) == nil
+}
+
+const (
+	// If the owner of an external service has this tag, the service is allowed to sync private code
+	TagAllowUserExternalServicePrivate = "AllowUserExternalServicePrivate"
+	// If the owner of an external service has this tag, the service is allowed to sync public code only
+	TagAllowUserExternalServicePublic = "AllowUserExternalServicePublic"
+)
+
+// SetTag adds (present=true) or removes (present=false) a tag from the given user's set of tags. An
+// error occurs if the user does not exist. Adding a duplicate tag or removing a nonexistent tag is
+// not an error.
+func (u *UserStore) SetTag(ctx context.Context, userID int32, tag string, present bool) error {
+	u.ensureStore()
+
+	var q *sqlf.Query
+	if present {
+		// Add tag.
+		q = sqlf.Sprintf(`UPDATE users SET tags=CASE WHEN NOT %s::text = ANY(tags) THEN (tags || %s::text) ELSE tags END WHERE id=%s`, tag, tag, userID)
+	} else {
+		// Remove tag.
+		q = sqlf.Sprintf(`UPDATE users SET tags=array_remove(tags, %s::text) WHERE id=%s`, tag, userID)
+	}
+
+	res, err := u.ExecResult(ctx, q)
+	if err != nil {
+		return err
+	}
+	nrows, err := res.RowsAffected()
+	if err != nil {
+		return err
+	}
+	if nrows == 0 {
+		return userNotFoundErr{args: []interface{}{userID}}
+	}
+	return nil
+}
+
+// HasTag reports whether the context actor has the given tag.
+// If not, it returns false and a nil error.
+func (u *UserStore) HasTag(ctx context.Context, userID int32, tag string) (bool, error) {
+	if Mocks.Users.HasTag != nil {
+		return Mocks.Users.HasTag(ctx, userID, tag)
+	}
+	u.ensureStore()
+
+	var tags []string
+	err := u.QueryRow(ctx, sqlf.Sprintf("SELECT tags FROM users WHERE id = %s", userID)).Scan(pq.Array(&tags))
+	if err != nil {
+		if err == sql.ErrNoRows {
+			return false, userNotFoundErr{[]interface{}{userID}}
+		}
+		return false, err
+	}
+
+	for _, t := range tags {
+		if t == tag {
+			return true, nil
+		}
+	}
+	return false, nil
 }
